+++ packages/ui/v2/core/index.ts
@@ -39,7 +39,6 @@ export { Portal } from "./Portal";
 export { Segment, SegmentOption } from "./Segment";
 export { default as SettingsToggle } from "./SettingsToggle";
 export { default as Shell } from "./Shell";
-export { Skeleton, SkeletonAvatar, SkeletonButton, SkeletonContainer, SkeletonText } from "./skeleton";
 export { default as Stepper } from "./Stepper";
 export { default as Swatch } from "./Swatch";
 export { default as Switch } from "./Switch";
